requirements.txt
# Core dependencies
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.1.3
numpy>=1.26.2
SQLAlchemy>=2.0.25

# Visualization
streamlit>=1.28.2
plotly>=5.18.0

# Testing
pytest>=7.4.3
PyYAML>=6.0.1
lark-parser>=0.12.0
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.2
packaging<24.0,>=16.8 